requirements.txt
discord.py>=2.3.0
apscheduler>=3.10.0
